# Conflicts:
#   commons/src/main/java/commons/Card.java

+++ client/src/main/java/client/scenes/TaskViewCtrl.java
@@ -1,17 +1,18 @@
 package client.scenes;
 
 import client.utils.ServerUtils;
-import com.google.inject.Stage;
 import commons.Board;
 import commons.Card;
 import commons.List;
 import javafx.fxml.FXML;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
-import javafx.scene.control.TextField;
+import javafx.scene.control.ButtonType;
 import javafx.scene.text.Text;
 
 import javax.inject.Inject;
+import java.util.Optional;
 
 public class TaskViewCtrl {
 
@@ -50,6 +51,12 @@ public class TaskViewCtrl {
     private Board boardCurr;
 
 
+    /**
+     * constructor
+     * @param server the current server
+     * @param mainCtrl a reference to the MainCtrl
+     * @param card the card that is being viewed
+     */
     @Inject
     public TaskViewCtrl(ServerUtils server, MainCtrl mainCtrl, Card card) {
         this.server = server;
@@ -58,6 +65,10 @@ public class TaskViewCtrl {
 
     }
 
+    /**
+     * show the information of the card in th UI
+     * @param card the card to show
+     */
     public void renderInfo(Card card){
         currCard = card;
         taskName.setText(card.name);
@@ -69,24 +80,48 @@ public class TaskViewCtrl {
         return;
     }
 
+    /**
+     * switches the scene to the taskEdit
+     */
     public void goEdit(){
         mainCtrl.switchEdit(currCard, boardCurr);
     }
-    public void goDelete(){
-        List listCurr=currCard.getList();
-        listCurr.cards.remove(currCard);
 
-        for(int i=0; i<boardCurr.lists.size(); i++){
-            if(boardCurr.lists.get(i).getID()==listCurr.getID()){
-                boardCurr.lists.set(i,listCurr);
+    /**
+     * throws a confirmation message for deleting the card and deletes the card
+     */
+    public void goDelete(){
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setHeaderText("Delete Task '" + currCard.getName() + "'?");
+        alert.setContentText("Are you sure you want to delete task '" + currCard.getName() +
+                "'?\nThis will permanently delete the task from the server.");
+
+        ButtonType delete = new ButtonType("Delete");
+        ButtonType cancel = new ButtonType("Cancel");
+        alert.getButtonTypes().setAll(delete, cancel);
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        if (result.get() == delete) {
+            List listCurr=currCard.getList();
+            listCurr.cards.remove(currCard);
+
+            for(int i=0; i<boardCurr.lists.size(); i++){
+                if(boardCurr.lists.get(i).getID()==listCurr.getID()){
+                    boardCurr.lists.set(i,listCurr);
+                }
             }
+
+            server.updateBoard(boardCurr);
+            server.deleteCard(currCard.id);
+            mainCtrl.switchDelete(currCard);
         }
 
-        server.updateBoard(boardCurr);
-        server.deleteCard(currCard.id);
-        mainCtrl.switchDelete(currCard);
     }
 
+    /**
+     * switches the scene to dashboard
+     */
     @FXML
     public void setDone(){
         mainCtrl.switchDashboard("LOL");
@@ -100,10 +135,18 @@ public class TaskViewCtrl {
         error.setText(err);
     }
 
+    /**
+     * sets the current list to a new one
+     * @param listCurr the new list
+     */
     public void setListCurr(List listCurr) {
         this.listCurr = listCurr;
     }
 
+    /**
+     * sets the current board to a new one
+     * @param boardCurr the new board
+     */
     public void setBoardCurr(Board boardCurr) {
         this.boardCurr = boardCurr;
     }
